Move current `Spartan` codec struct into `subspace-codec` crate that will later be refactored into proper Subspace codec
