+++ crates/pallet-spartan/src/mock.rs
@@ -27,9 +27,7 @@ use ring::{digest, hmac};
 use schnorrkel::{Keypair, PublicKey};
 use sp_consensus_poc::digests::{PreDigest, Solution};
 use sp_consensus_poc::Slot;
-use sp_consensus_spartan::spartan::{
-    Piece, Tag, ENCODE_ROUNDS, GENESIS_PIECE_SEED, PIECE_SIZE, PRIME_SIZE_BYTES, SIGNING_CONTEXT,
-};
+use sp_consensus_spartan::spartan::{Tag, SIGNING_CONTEXT};
 use sp_core::sr25519::Pair;
 use sp_core::{Pair as PairTrait, Public, H256};
 use sp_io;
@@ -39,8 +37,7 @@ use sp_runtime::{
     Perbill,
 };
 use std::convert::TryInto;
-use std::io::Write;
-use subspace_core_primitives::{LastArchivedBlock, RootBlock, Sha256Hash};
+use subspace_core_primitives::{LastArchivedBlock, Piece, RootBlock, Sha256Hash, PRIME_SIZE};
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
@@ -181,12 +178,11 @@ pub fn go_to_block(keypair: &Keypair, block: u64, slot: u64) {
         System::parent_hash()
     };
 
-    let spartan =
-        sp_consensus_spartan::spartan::Spartan::new(genesis_piece_from_seed(GENESIS_PIECE_SEED));
+    let spartan = subspace_codec::Spartan::new();
     let public_key_hash = hash_public_key(&keypair.public);
     let ctx = schnorrkel::context::signing_context(SIGNING_CONTEXT);
     let nonce = 0;
-    let encoding: Piece = spartan.encode(public_key_hash, nonce, ENCODE_ROUNDS);
+    let encoding: Piece = spartan.encode(public_key_hash, nonce);
     let tag: Tag = create_tag(&encoding, &Spartan::salt().to_le_bytes());
 
     let pre_digest = make_pre_digest(
@@ -205,20 +201,10 @@ pub fn go_to_block(keypair: &Keypair, block: u64, slot: u64) {
     Spartan::on_initialize(block);
 }
 
-fn genesis_piece_from_seed(seed: &str) -> Piece {
-    let mut piece = [0u8; PIECE_SIZE];
-    let mut input = seed.as_bytes().to_vec();
-    for mut chunk in piece.chunks_mut(digest::SHA256.output_len) {
-        input = digest::digest(&digest::SHA256, &input).as_ref().to_vec();
-        chunk.write_all(input.as_ref()).unwrap();
-    }
-    piece
-}
-
-fn hash_public_key(public_key: &PublicKey) -> [u8; PRIME_SIZE_BYTES] {
-    let mut array = [0u8; PRIME_SIZE_BYTES];
+fn hash_public_key(public_key: &PublicKey) -> [u8; PRIME_SIZE] {
+    let mut array = [0u8; PRIME_SIZE];
     let hash = digest::digest(&digest::SHA256, public_key.as_ref());
-    array.copy_from_slice(&hash.as_ref()[..PRIME_SIZE_BYTES]);
+    array.copy_from_slice(&hash.as_ref()[..PRIME_SIZE]);
     array
 }
