+++ crates/sc-consensus-poc/src/tests.rs
@@ -35,18 +35,16 @@ use sc_network_test::{
 use schnorrkel::{Keypair, PublicKey};
 use sp_consensus::{AlwaysCanAuthor, DisableProofRecording, NoNetwork as DummyOracle, Proposal};
 use sp_consensus_poc::{inherents::InherentDataProvider, Slot};
-use sp_consensus_spartan::spartan::{
-    Piece, Tag, ENCODE_ROUNDS, GENESIS_PIECE_SEED, PIECE_SIZE, PRIME_SIZE_BYTES,
-};
+use sp_consensus_spartan::spartan::Tag;
 use sp_core::Public;
 use sp_runtime::{
     generic::DigestItem,
     traits::{Block as BlockT, DigestFor},
 };
 use sp_timestamp::InherentDataProvider as TimestampInherentDataProvider;
-use std::io::Write;
 // use std::sync::mpsc;
 use std::{cell::RefCell, task::Poll, time::Duration};
+use subspace_core_primitives::PRIME_SIZE;
 use substrate_test_runtime::{Block as TestBlock, Hash};
 
 type Item = DigestItem<Hash>;
@@ -382,7 +380,7 @@ impl TestNetFactory for PoCTestNet {
                 epoch_changes: data.link.epoch_changes.clone(),
                 can_author_with: AlwaysCanAuthor,
                 telemetry: None,
-                spartan: Spartan::default(),
+                spartan: Spartan::new(),
                 signing_context: schnorrkel::context::signing_context(SIGNING_CONTEXT),
             },
             mutator: MUTATOR.with(|m| m.borrow().clone()),
@@ -506,14 +504,12 @@ fn run_one_test(mutator: impl Fn(&mut TestHeader, Stage) + Send + Sync + 'static
 
         let mut new_slot_notification_stream = data.link.new_slot_notification_stream().subscribe();
         let poc_farmer = async move {
-            let spartan = sp_consensus_spartan::spartan::Spartan::new(genesis_piece_from_seed(
-                GENESIS_PIECE_SEED,
-            ));
+            let spartan = Spartan::new();
             let keypair = Keypair::generate();
             let public_key_hash = hash_public_key(&keypair.public);
             let ctx = schnorrkel::context::signing_context(SIGNING_CONTEXT);
             let nonce = 0;
-            let encoding: Piece = spartan.encode(public_key_hash, nonce, ENCODE_ROUNDS);
+            let encoding: Piece = spartan.encode(public_key_hash, nonce);
 
             while let Some(NewSlotNotification {
                 new_slot_info,
@@ -561,20 +557,10 @@ fn run_one_test(mutator: impl Fn(&mut TestHeader, Stage) + Send + Sync + 'static
     ));
 }
 
-fn genesis_piece_from_seed(seed: &str) -> Piece {
-    let mut piece = [0u8; PIECE_SIZE];
-    let mut input = seed.as_bytes().to_vec();
-    for mut chunk in piece.chunks_mut(digest::SHA256.output_len) {
-        input = digest::digest(&digest::SHA256, &input).as_ref().to_vec();
-        chunk.write_all(input.as_ref()).unwrap();
-    }
-    piece
-}
-
-fn hash_public_key(public_key: &PublicKey) -> [u8; PRIME_SIZE_BYTES] {
-    let mut array = [0u8; PRIME_SIZE_BYTES];
+fn hash_public_key(public_key: &PublicKey) -> [u8; PRIME_SIZE] {
+    let mut array = [0u8; PRIME_SIZE];
     let hash = digest::digest(&digest::SHA256, public_key.as_ref());
-    array.copy_from_slice(&hash.as_ref()[..PRIME_SIZE_BYTES]);
+    array.copy_from_slice(&hash.as_ref()[..PRIME_SIZE]);
     array
 }
