+++ crates/spartan-farmer/src/main.rs
@@ -18,7 +18,6 @@
 mod commands;
 mod crypto;
 mod plot;
-mod spartan;
 mod utils;
 
 use async_std::task;
@@ -33,8 +32,7 @@ use tokio::runtime::Runtime;
 type Tag = [u8; 8];
 type Salt = [u8; 8];
 
-const PRIME_SIZE_BYTES: usize = 32;
-const ENCODE_ROUNDS: usize = 1;
+const PRIME_SIZE: usize = 32;
 const SIGNING_CONTEXT: &[u8] = b"FARMER";
 const BATCH_SIZE: u64 = (16 * 1024 * 1024 / PIECE_SIZE) as u64;
 const CUDA_BATCH_SIZE: u64 = (32 * 1024) as u64;
