+++ crates/subspace-codec/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "subspace-codec"
+description = "Encoder for the Subspace Network Blockchain based on the SLOTH permutation"
+license = "Apache-2.0"
+version = "0.1.0"
+authors = ["Nazar Mokrynskyi <nazar@mokrynskyi.com>"]
+edition = "2018"
+include = [
+    "/src",
+    "/Cargo.toml",
+    "/LICENSE",
+    "/README.md"
+]
+
+[dependencies]
+ring = "0.16.20"
+sloth256-189 = "0.2.2"
+
+[dependencies.subspace-core-primitives]
+version = "0.1.0"
+path = "../subspace-core-primitives"
+
+[dev-dependencies]
+rand = "0.8.4"
+
+[features]
+default = []
+# Compile with CUDA support and use it if compatible GPU is available
+cuda = [
+    "sloth256-189/cuda",
+]
